Add files via upload
So this is my Sovastolink project that i've been working on for about a month now i would like to say that ther are still things that need to be done like to make the playsound code to work and to make the if statement code to work if you see in my a sovasotlink.yml file i have some goals that i've set for myself. This project is going to be updated untill i start making a new project. So if anyone has any ideas on making new features to this program pls tell me i would really like to know so i could try my best to add them. If you only download the .exe program there is a chance where your windows defender might go off and not open the program, but if that happens to you pls run the code on your computer and it should work that way.